Capture/Replay: Clean up EGL capture.
This switches the EGL capture types to ANGLE-casted pointers since
that's what we receive in the capture layer. Note that even if the
capture layer were used as a pure layer, not an EGL implementation,
we'd still have these types for state tracking.

This also prefixes each EGL class in the entry points with the egl
namespace for consistency and for simplifying the ParamType code.

Required changing to non-const gl::Context * in a few places. Also
changes the gSurfaceMap to be indexed by the raw pointer value,
which cleans up the code somewhat.

Bug: angleproject:4035
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/3949910
Commit-Queue: Jamie Madill <jmadill@chromium.org>

+++ src/libANGLE/capture/trace_fixture.h
@@ -113,7 +113,7 @@ extern SyncResourceMap gSyncMap;
 using ContextMap = std::unordered_map<uint32_t, EGLContext>;
 extern ContextMap gContextMap;
 
-using SurfaceMap = std::unordered_map<uint32_t, EGLSurface>;
+using SurfaceMap = std::unordered_map<uintptr_t, EGLSurface>;
 extern SurfaceMap gSurfaceMap;
 
 void UpdateClientArrayPointer(int arrayIndex, const void *data, uint64_t size);
